chore: preserve child compiler plugin version from git history
